add linters to post_gen_proj
