+++ CONTRIBUTING.MD
@@ -1,6 +1,7 @@
 # Troubleshooting
 
 ## non-Python Package dependencies
+
 `Cookiecutter-relecloud` builds instances that include support for [DevContainers](https://containers.dev) and deployments via [Azure CLI](https://aka.ms/azure-cli) or [Azure Developer CLI](https://aka.ms/azd).
 
 You will need to ensure all required products are installed and up to date.
@@ -32,7 +33,9 @@ curl -fsSL https://aka.ms/install-azd.sh | bash
 ## Testing
 
 ### Creating Tests
+
 We have two sets of tests:
+
 1. Generation tests - Tests that verify project generation is accurate
 2. Application tests - Tests that are created by cookiecutter that are ran by the generated project
 
@@ -40,7 +43,7 @@ Both sets of tests are created using [Pytest](https://pytest.org).
 
 ### Generation tests
 
-The generation tests are located at `/tests/` 
+The generation tests are located at `/tests/`
 
 Generation tests use [pytest-cookies](https://github.com/hackebrot/pytest-cookies) to generate an iteration of each project and run a series of tests.
 
@@ -60,7 +63,6 @@ Application tests use [playwright](https://playwright.dev/python/) and [pytest-p
 playwright install --with-deps
 ```
 
-
 ### Running Tests
 
 Our goal is to ensure that the template is always up to date with the latest versions of the tools we use. Tests are run using [GitHub Actions](https://github.com/kjaymiller/cookiecutter-relecloud/actions). You should also run the tests locally before submitting a PR.
@@ -68,6 +70,7 @@ Our goal is to ensure that the template is always up to date with the latest ver
 There are plenty of tests our repo. The majority of tests are focused on testing deployment scendarios. These cannot be tested without creating a subprocess. These tests are relatively slow and should not be ran on every commit. Instead, they are ran when a PR is submitted.
 
 ## Adding a New Framework
+
 If you would like to add a new framework there are a few things that you need to consider.
 
 - What database technologies is the framework compatible with?
@@ -77,6 +80,7 @@ If you would like to add a new framework there are a few things that you need to
 - What are the best practices for the framework on Azure?
 
 ### Database Technology Support
+
 We currently support the following database technologies:
     - PostgreSQL
     - MongoDB
@@ -85,6 +89,36 @@ Your framework should be compatible with at least one of these technologies. If 
 
 ## Adding a New Database
 
+## Post Build Work
+
+### Linting Files
+
+Generating files often have issues with whitespace among other things. To ensure that our generated code is a clean as possible, we run a linter on the generated files. You can add your linter function in `{{cookiecutter.__src_folder_name}}/hooks/linters.py`. If you are trying to add a linting function, be sure to call it in the `lint()` function.
+
+```python
+# {{cookiecutter.__src_folder_name}}/hooks/linters.py
+
+def my_new_linter() -> None:
+    """Linting function for file type"""
+    # Do something
+    
+... # Other linters
+
+def lint() -> None:
+    ... # other linters
+    my_new_linter()
+```
+
+```python
+```
+
+### Removing Unused Files
+
+If you are adding functionality that requires specific files for that resource to work, you will need to add a post build step to remove the files that are not needed if that resource is not selected.
+
+You can add your removal function in `{{cookiecutter.__src_folder_name}}/hooks/removers.py` and add a check for it in the `check for files()` function.
+
+```python
 
 ## Testing Frameworks and Databases
 In our `tests/conftest.py` we have a fixture style we refer to as `bakery`. These session fixtures are used to generate a project and run tests against it.
